+++ internal/provider/testdata/test-action-dir/action.tf
@@ -1,0 +1,5 @@
+action "scaffolding_example" "example1" {
+  config {
+    required_attr = "value-1"
+  }
+}