+++ internal/provider/validate.go
@@ -26,8 +26,8 @@ const (
 	FileExtensionMarkdown     = `.markdown`
 	FileExtensionMd           = `.md`
 
-	DocumentationGlobPattern    = `{docs/index.*,docs/{,cdktf/}{data-sources,ephemeral-resources,guides,resources,functions}/**/*,website/docs/**/*}`
-	DocumentationDirGlobPattern = `{docs/{,cdktf/}{data-sources,ephemeral-resources,guides,resources,functions}{,/*},website/docs/**/*}`
+	DocumentationGlobPattern    = `{docs/index.*,docs/{,cdktf/}{actions,data-sources,ephemeral-resources,guides,resources,functions}/**/*,website/docs/**/*}`
+	DocumentationDirGlobPattern = `{docs/{,cdktf/}{actions,data-sources,ephemeral-resources,guides,resources,functions}{,/*},website/docs/**/*}`
 )
 
 var ValidLegacyFileExtensions = []string{
@@ -314,6 +314,10 @@ func (v *validator) validateStaticDocs() error {
 		ephemeralResourceFiles, _ := fs.ReadDir(v.providerFS, dir+"/ephemeral-resources")
 		mismatchOpt.EphemeralResourceEntries = ephemeralResourceFiles
 	}
+	if dirExists(v.providerFS, dir+"/actions") {
+		actionFiles, _ := fs.ReadDir(v.providerFS, dir+"/actions")
+		mismatchOpt.ActionEntries = actionFiles
+	}
 
 	v.logger.infof("running file mismatch check")
 	if err := check.NewFileMismatchCheck(mismatchOpt).Run(); err != nil {
@@ -400,6 +404,10 @@ func (v *validator) validateLegacyWebsite() error {
 	if dirExists(v.providerFS, dir+"/ephemeral-resources") {
 		ephemeralResourceFiles, _ := fs.ReadDir(v.providerFS, dir+"/ephemeral-resources")
 		mismatchOpt.EphemeralResourceEntries = ephemeralResourceFiles
+	}
+	if dirExists(v.providerFS, dir+"/actions") {
+		actionFiles, _ := fs.ReadDir(v.providerFS, dir+"/actions")
+		mismatchOpt.ActionEntries = actionFiles
 	}
 
 	v.logger.infof("running file mismatch check")